refacto: move group type to types
